# src/swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "1.0.0"
  title: "Hookah App"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
basePath: "/"
...
schemes:
  - "http"
  - "https"
paths:
...
  /api/balances:
    post:
      tags:
        - "balances"
      symmary: ""
      descrpition: ""
      operationId: "createBalance"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: hookah-jwt
          in: header
          description: an authorization header
          type: string
        - in: "body"
          name: "body"
          description: "The balance that needs to create"
          required: true
          type: "string"
          schema:
            $ref: "#/definitions/BalanceCreate"
      responses:
        200:
          schema:
            type: "object"
            items:
              $ref: "#/definitions/BalanceCreate"
        400:
          description: "Invalid status value"
    get:
      tags:
      - "balances"
      summary: ""
      description: ""
      operationId: "getAllBalances"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Balance"
        400:
          description: "Invalid status value"
  /api/balances/{balanceId}:
    patch:
      tags:
      - "balances"
      summary: ""
      description: ""
      operationId: "get(debit|credit)"
      parameters:
      - name: "balanceId"
        in: "path"
        description: "The balance id that needs to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "The balance id that needs to be updated"
        required: true
        type: "string"
        schema:
          $ref: "#/definitions/BalanceAction"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          schema:
            type: "object"
            items:
              $ref: "#/definitions/Balance"
        400:
          description: "Invalid status value"
...
definitions:
...
  Balance:
    type: "object"
    properties:
      email:
        type: "string"
        description: "TBD"
      amount:
        type: "number"
      message:
        type: "string"
      createdAt:
        type: "string"
  BalanceAction:
    type: "object"
    properties:
      action:
        type: "string"
        description: "TBD"
        enum:
        - "debit"
        - "credit"
      amount:
        type: "number"
...
  BalanceCreate:
    type: "object"
    properties:
      userId:
        type: "string"
        description: "TBD"
      amount:
        type: "number"
      message:
        type: "string"